+++ siaw_DBContext/Models_Extra/sldosItemCompleto.cs
@@ -3,6 +3,6 @@
     public class sldosItemCompleto
     {
         public string descripcion { get; set; }
-        public float valor { get; set; }
+        public double valor { get; set; }
     }
 }
